update restful service

+++ restful-service/src/main/java/io/cygnus/service/dto/LastPrice.java
@@ -15,6 +15,6 @@ public class LastPrice {
 
 	private String instrumentCode;
 
-	private long lastPrice;
+	private double lastPrice;
 
 }
